// src/app/enums/form-controls.enum.ts
export enum UpdateFormControlsEnum {
  JOB_NUMBER = 'jobNumber',
  PRIORITY = 'priority',
  TYPE = 'type',
  LINE = 'line',
  MACHINE_CODE = 'machineCode',
  MACHINE_NAME = 'machineName',
  OCC_DATE = 'occDate',
  OCC_TIME = 'occTime',
  TR = 'tr',
  JOB_DESC = 'jobDesc',
  SHOP = 'shop',
  JOB_INITIATED_BY = 'shopInitiatedBy',
  USER_ID = 'userId'
}